@@ -1,4 +1,4 @@
-﻿using System.Drawing;
+using System;
 
 namespace OpenMacroBoard.SDK
 {
@@ -8,47 +8,13 @@ namespace OpenMacroBoard.SDK
     public interface IKeyBitmapDataAccess
     {
         /// <summary>
-        /// Gets a value indicating wheter the underlying byte array is null
+        /// Gets a value indicating whether the underlying byte array is null.
         /// </summary>
-        /// <value></value>
-        bool IsNull { get; }
+        bool IsEmpty { get; }
 
         /// <summary>
-        /// Gets the stride for the bitmap data
+        /// Gets the underlying image data in unaligned Bgr24 format (stride = width * 3).
         /// </summary>
-        int Stride { get; }
-
-        /// <summary>
-        /// Gets the length of the bitmap data array
-        /// </summary>
-        int DataLength { get; }
-
-        /// <summary>
-        /// Copies <paramref name="length"/> number of bytes from the bitmap data array to a given array.
-        /// </summary>
-        /// <param name="targetArray">Target array</param>
-        /// <param name="targetIndex">Index of first byte in <paramref name="targetArray"/></param>
-        /// <param name="startIndex">Index of first byte in bitmap data array</param>
-        /// <param name="length">Number of bytes to copy</param>
-        void CopyData(byte[] targetArray, int targetIndex, int startIndex, int length);
-
-        /// <summary>
-        /// Copies the bitmap data array to a given array.
-        /// </summary>
-        /// <param name="targetArray">Target array</param>
-        /// <param name="targetIndex">Index of first byte in <paramref name="targetArray"/></param>
-        void CopyData(byte[] targetArray, int targetIndex);
-
-        /// <summary>
-        /// Creates a copy if the internal bitmap data array
-        /// </summary>
-        /// <returns>raw bgr24 bitmap array</returns>
-        byte[] CopyData();
-
-        /// <summary>
-        /// Creates a new bitmap based on the internal raw pixeldata.
-        /// </summary>
-        /// <returns>Returns the generated bitmap.</returns>
-        Bitmap GetBitmap();
+        ReadOnlySpan<byte> GetData();
     }
 }
