@@ -1,4 +1,6 @@
-﻿namespace OpenMacroBoard.SDK
+using System;
+
+namespace OpenMacroBoard.SDK
 {
     /// <summary>
     /// A bunch of extensions to clear all keys, or set a single <see cref="KeyBitmap"/> to all keys.
@@ -8,13 +10,12 @@
         /// <summary>
         /// Sets a background image for all keys
         /// </summary>
-        /// <param name="board"></param>
-        /// <param name="bitmap"></param>
+        /// <exception cref="ArgumentNullException">The provided board is null.</exception>
         public static void SetKeyBitmap(this IMacroBoard board, KeyBitmap bitmap)
         {
             if (board is null)
             {
-                throw new System.ArgumentNullException(nameof(board));
+                throw new ArgumentNullException(nameof(board));
             }
 
             for (var i = 0; i < board.Keys.Count; i++)
@@ -26,13 +27,12 @@
         /// <summary>
         /// Sets background to black for a given key
         /// </summary>
-        /// <param name="board"></param>
-        /// <param name="keyId"></param>
+        /// <exception cref="ArgumentNullException">The provided board is null.</exception>
         public static void ClearKey(this IMacroBoard board, int keyId)
         {
             if (board is null)
             {
-                throw new System.ArgumentNullException(nameof(board));
+                throw new ArgumentNullException(nameof(board));
             }
 
             board.SetKeyBitmap(keyId, KeyBitmap.Black);
@@ -41,7 +41,6 @@
         /// <summary>
         /// Sets background to black for all given keys
         /// </summary>
-        /// <param name="board"></param>
         public static void ClearKeys(this IMacroBoard board)
         {
             board.SetKeyBitmap(KeyBitmap.Black);
